refactor(core): remove shared.ts, consolidate tap/doubletap in gesture.ts
shared.ts was just a coordinator renamed — it baked in TAP_THRESHOLD,
tap-specific pointer listeners, and a PointerTapHandler type. Nothing
in it was reusable by a future swipe gesture.

Collapse back into gesture.ts as self-contained tap/doubletap module.
When swipe arrives it gets its own file with its own state — no fake
shared abstraction needed until then.
