+++ packages/core/src/dom/index.ts
@@ -1,7 +1,6 @@
 export * from './feature';
 export * from './gesture/gesture';
 export * from './gesture/region';
-export * from './gesture/shared';
 export * from './hotkey/actions';
 export * from './hotkey/aria';
 export * from './hotkey/coordinator';
